+++ app/src/lib/dispatcher/dispatcher.ts
@@ -23,7 +23,7 @@ import { fatalError } from '../fatal-error'
 import { structuralEquals } from '../equality'
 import { isGitOnPath } from '../open-shell'
 import { uuid } from '../uuid'
-import { URLActionType, IOpenRepositoryArgs } from '../parse-url'
+import { URLActionType, IOpenRepositoryArgs } from '../parse-app-url'
 import { requestAuthenticatedUser, resolveOAuthRequest, rejectOAuthRequest } from '../../lib/oauth'
 import { validatedRepositoryPath } from './validated-repository-path'
